Handle null values returned from `getCTM`
According the the [SVG standard][] the `getCTM` method may return
`null`. This is *supposed* to only happen when the SVG element is not
attached to a document, however I'm observing that it can be returned
even when the SVG is attached to the document in Firefox. To hack
around this scenario, we simply use un-transformed points when given
a `null` CTM, which seems to work correctly.

Fixes https://github.com/nornagon/saxi/issues/34

[SVG standard]: https://www.w3.org/TR/SVG/types.html#__svg__SVGGraphicsElement__getCTM